requirements.txt
# Google GenAI / ADK
google-genai>=0.0.37
google-adk>=0.0.15

# Google Cloud BigQuery
google-cloud-bigquery>=3.11.0

# Async & HTTP
aiohttp>=3.8.5
httpx>=0.26.0

# Data handling & analysis
pandas>=2.1.0
numpy>=1.25.0

# YAML / JSON utilities
PyYAML>=6.0
orjson>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Logging & debugging
loguru>=0.7.0

# Optional: Data Science tools for BQML analysis
scikit-learn>=1.3.2
matplotlib>=3.8.0
seaborn>=0.12.2
